Consolidate App\ShellCommand and App\Probe into App\Slave\Task
