+++ src/Command/ProbeDispatcherCommand.php
@@ -9,8 +9,8 @@ use App\DependencyInjection\SlaveConfiguration;
 use App\DependencyInjection\StatsManager;
 use App\DependencyInjection\WorkerManager;
 use App\Instruction\Instruction;
-use App\ShellCommand\GetConfigHttpWorkerCommand;
-use App\ShellCommand\PostStatsHttpWorkerCommand;
+use App\Slave\Task\GetConfigHttpWorkerCommand;
+use App\Slave\Task\PostStatsHttpWorkerCommand;
 use Exception;
 use Psr\Log\LoggerInterface;
 use React\EventLoop\Factory;
