+++ src/DependencyInjection/Compiler/CommandPass.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\DependencyInjection\Compiler;
 
-use App\ShellCommand\CommandFactory;
+use App\Slave\Task\CommandFactory;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
