+++ src/DependencyInjection/StatsManager.php
@@ -2,9 +2,9 @@
 
 namespace App\DependencyInjection;
 
-use App\ShellCommand\GetConfigHttpWorkerCommand;
-use App\ShellCommand\PostResultsHttpWorkerCommand;
-use App\ShellCommand\PostStatsHttpWorkerCommand;
+use App\Slave\Task\GetConfigHttpWorkerCommand;
+use App\Slave\Task\PostResultsHttpWorkerCommand;
+use App\Slave\Task\PostStatsHttpWorkerCommand;
 use Psr\Log\LoggerInterface;
 
 class StatsManager
