+++ src/Slave/Task/PostResultsHttpWorkerCommand.php
@@ -2,15 +2,14 @@
 
 declare(strict_types=1);
 
-namespace App\ShellCommand;
+namespace App\Slave\Task;
 
 use App\Client\FirepingClient;
 use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Exception\RequestException;
 use Psr\Log\LoggerInterface;
-use stdClass;
 
-class PostStatsHttpWorkerCommand implements CommandInterface
+class PostResultsHttpWorkerCommand implements CommandInterface
 {
     /**
      * @var LoggerInterface
@@ -33,7 +32,7 @@ class PostStatsHttpWorkerCommand implements CommandInterface
     public function execute(): array
     {
         $startedAt = microtime(true);
-        $this->logger->info(sprintf('worker: publishing stats to master (%d bytes)', strlen(serialize($this->body))));
+        $this->logger->info(sprintf('worker: publishing results (%d bytes)', strlen(serialize($this->body))));
         try {
             $response = $this->client->request($this->method, $this->endpoint, ['json' => $this->body]);
 
@@ -56,8 +55,8 @@ class PostStatsHttpWorkerCommand implements CommandInterface
     public function setArgs(array $args): void
     {
         $this->method = $args['method'] ?? 'POST';
-        $this->endpoint = sprintf('/api/slaves/%s/stats', $_ENV['SLAVE_NAME']);
-        $this->body = $args['body'] ?? new stdClass();
+        $this->endpoint = sprintf('/api/slaves/%s/result', $_ENV['SLAVE_NAME']);
+        $this->body = $args['body'] ?? new \stdClass();
     }
 
     public function getType(): string
